Tools for optimizing distillation

Tools for optimizing distillation

The following functions implement the various methods to optimize entanglement distillation described in arXiv:XXX.

EntanglementDist.pptRelax — Function.

(problem, F, p_succ) = pptRelax(rho,n,k,delta, verbose, eps, max_iters)

Implements the PPT relaxation computing an upper bound on the fidelity achievable using realistic operations for a given input state and fixed succes probability.

Inputs:

rho quantum state to be distilled

n number of qubits on one side, assuming dimensions nA=nB =2^n in the input.

k desired output dimension of the maximally entangled state

delta desired success probability

verbose flag indicating whether the SCS solver should be verbose (default false)

eps desired precision (default 1e-4) in SCS solver

max_iters maximum number of iterations in SCS solver (default 2e4)

Outputs:

problem SDP object from Convex.jl allowing further processing.

F Fidelity achieved

p_succ success probability attained in SDP (should equal delta)

(problem, F, p_succ) = pptRelax(rho,nA, nB, k,delta, verbose, eps, max_iters)

Implements the PPT relaxation computing an upper bound on the fidelity achievable using realistic operations for a given input state and fixed succes probability.

Inputs:

rho quantum state to be distilled on A and B

nA dimension of the A system

nB dimension of the B system

k desired output dimension of the maximally entangled state

delta desired success probability

verbose flag indicating whether the SCS solver should be verbose (default true)

eps desired precision (default 1e-4) in SCS solver

max_iters maximum number of iterations in SCS solver (default 2e4)

Outputs:

problem SDP object from Convex.jl allowing further processing.

F Fidelity achieved

p_succ success probability attained in SDP (should equal delta)

EntanglementDist.pptRelaxCopies — Function.

(problem, F, p) = pptRelaxCopies(rho, n, nA, nB, k, delta, verbose, eps, max_iters)

Calls pptRelax for an input state of the form rho&^(tensor n), reordered appropriately.

EntanglementDist.pptRelax1Ext — Function.

(problem, F, psucc) = pptRelax1Ext(rhoAB, n, k, delta, verbose=true/false, eps=1e^-4)

Implements the PPT relaxation plus 1 symmetric extension for distillable entanglement with fixed desired success probability.

Inputs:

rhoAB quantum state to be distilled

n number of qubits on one side, assuming dimensions nA=nB =2^n in the input.

k local dimension of the maximally entangled output state

delta desired success probability

Outputs:

problem problem object given by Convex,

F fidelity bound

psucc success probability actually attained (should equal delta up to numerical imprecisions)

(problem, F, psucc) = pptRelax1Ext(rhoAB, nA, nB, k, delta, verbose=true/false, eps=1e^-4)

Implements the PPT relaxation plus 1 symmetric extension for distillable entanglement with fixed desired success probability.

Inputs:

rhoAB quantum state to be distilled

nA dimension of the A system

nB dimension of the B system

k local dimension of the maximally entangled output state

delta desired success probability

Outputs:

problem problem object given by Convex,

F fidelity bound

psucc success probability actually attained (should equal delta up to numerical imprecisions)

EntanglementDist.pptRelax1ExtCopies — Function.

(problem, F, p) = pptRelax1ExtCopies(rho, n, nA, nB, k, delta, verbose, eps, max_iters)

Calls pptRelax1Ext for an input state of the form rho&^(tensor n), reordered appropriately.

EntanglementDist.pptRelax2Ext — Function.

(problem, F, psucc) = pptRelax2Ext(rhoAB, n, k, delta, verbose=true/false, eps=1e^-4)

Implements the PPT relaxation plus 2 extensions for distillable entanglement with fixed desired success probability.

Inputs:

rhoAB quantum state to be distilled

n number of qubits on one side, assuming dimensions nA=nB =2^n in the input.

k local dimension of the maximally entangled output state

delta desired success probability

Outputs:

problem problem object given by Convex,

F fidelity bound

psucc success probability actually attained (should equal delta up to numerical imprecisions)

(problem, F, psucc) = pptRelax2Ext(rhoAB, nA, nB, k, delta, verbose=true/false, eps=1e^-4)

Implements the PPT relaxation plus 2 extensions for distillable entanglement with fixed desired success probability.

Inputs:

rhoAB quantum state to be distilled

nA dimension of the A system

nB dimension of the B system

k local dimension of the maximally entangled output state

delta desired success probability

Outputs:

problem problem object given by Convex,

F fidelity bound

psucc success probability actually attained (should equal delta up to numerical imprecisions)

EntanglementDist.pptRelax2ExtCopies — Function.

(problem, F, p) = pptRelax2ExtCopies(rho, n, nA, nB, k, delta, verbose, eps, max_iters)

Calls pptRelax2Ext for an input state of the form rho&^(tensor n), reordered appropriately.

EntanglementDist.seesaw — Function.

(CA,CB, F, psucc) = seesaw(rhoAB, nA, nB, k, startCA, startCB) or (CA,CB, F, psucc) = seesaw(rhoAB, nA, nB, k, startCA, startCB, P) or (CA,CB, F, psucc) = seesaw(rhoAB, nA, nB, k, startCA, startCB, P, max_iter) or (CA,CB, F, psucc) = seesaw(rhoAB, nA, nB, k, startCA, startCB, P, max_iter, delta)

Performs a seesaw optimization searching for a good scheme to distill rhoAB into a maximally entangled state of local dimension k. Good here means that we want to maximize the fidelity of the output state to the maximally entangled state, where the success probability should not fall below the existing one - or, if supplied, not below the minimum success probability delta.

Alice's map is assumed to be from A -> hatA, F_A, and Bob's from B -> hatB, F_B, where F_A and F_B are the flag registered that yield success or failure. After the map P is measured on the flag registeres to determine success where P is the projector determining success. Example P=|11><11| on the flag registers F_A and F_B.

Input:

rhoAB state to be distilled

nA dimension of A

nB dimension of B

k local dimension of the desired maximally entangled output state

startCA choi state of Alice's map to start from (ordering of systems: hatA (output state), F_A (flag register), A (input))

startCB choi state of Bob's map to start from (ordering of systems: hatB (output state), F_B (flag register), B (input))

P (optional) projector onto success: Default is |11><11|

max_iter (optional) maximum number of iterations (default 10)

delta (optional) minimum acceptable success probability.

Output

CA new map for Alice as choi state

CB new map for Bob as choi state

F fidelity obtained

psucc success probability attained