Tools for creating and manipulating quantum states

Within our package we provide a number of useful functions for dealing with quantum states.

EntanglementDist.isQuantumState — Function.

b = isQuantumState(rho) or b = isQuantumState(rho,prec)

Checks whether the input matrix rho is a valid quantum state, that is, a matrix that is positive semidefinite and has trace 1. prec is the precision that small eigenvalues are dealt with: If the input matrix has a negative eigenvalue that is no smaller than -prec, then this eigenvalue is considered zero.

Returns true/false.

EntanglementDist.isHermitian — Function.

'b = isHermitian(rho)' or 'b = is Hermitian(rho, prec)'

Checks whether the input matrix rho is a Hermitian matrix. Entries are compares up to precision prec on average.

Returns true/false.

EntanglementDist.isUnitary — Function.

'b = isUnitary(U)'

Checks whether the input matrix is unitary.

Returns true/false.

EntanglementDist.eVec — Function.

vec = eVec(d, j)

Returns a sparse vector with all 0's except at the position j.

EntanglementDist.maxEnt — Function.

vec = maxEnt(d)

Returns a projector (density matrix) onto the maximally entangled state of a given dimension d.

EntanglementDist.maxEntVec — Function.

vec = maxEntVec(d)

Returns the maximally entangled state of a given dimension d as a vector.

EntanglementDist.copies — Function.

rhoOut = copies(rho,n)

Returns rho^(tensor n), i.e., n copies of the input rho

EntanglementDist.sortAB — Function.

rhoSorted = sortAB(rhoBig, d, n)

Given a state of the form rhoBig = rhoAB^(tensor n), rearrange, so all A parts are first, followed by all B parts. d is the dimension of A, assumed to be the same as for B. n is the number of copies.

EntanglementDist.entFidelity — Function.

F = entFidelity(rho)

Returns the fidelity with a maximally entangled state of equal dimension.

EntanglementDist.partialtrace — Function.

rhoOut = partialtrace(rhoIn, sysNum, dimVec)

Computes the partial trace of rhoIn, tracing out system number sysNum, where the dimensions of the systems are specified in the vector dimVec.

Example: rhoOut = partialtrace(epr, 2, [2 2])

Traces out the second qubit of a 2x2 state, here the EPR pair.

EntanglementDist.partialtranspose — Function.

Apply the partial tanspose to multiple sytems.

Apply partial transpose on systems if naive is false. Otherwise do the naive partial transpose in blocks of l × l. This is equivalent to applying the partial transepose to the last system if that particular system has dimension l.