Known distillation/filtering schemes

Known distillation/filtering schemes

Compute how well several known entanglement distillation schemes perform on specific input states. We also implement a simple filtering scheme and provide function for computing the performance of general filtering schemes given the relevant measurement operators. The functions computing the choi states of said filtering protocols can be used to test the performance of the seesaw method.

EntanglementDist.DEJMPSParam — Function.

(F, p_succ) = DEJMPSParam(rho)

Determines the performance of DEJMPS for two copies of the input state rho (consisting of two qubits) and returns the fidelity and success probability achieved. The procedure optimises over local single qubit rotations that permute the bell diagonal coefficients before the CNOTs to achieve the highest output fidelity. returns (F_out, p_succ)

EntanglementDist.BBPSSWParam — Function.

(F, p_succ) = BBPSSWParam(F)

Determines the performance of BBPSSW for two copies of the input state (consisting of two qubits) with fidelity F each. returns (F_out, p_succ)

(F, p_succ) = BBPSSWParam(rho)

Determines the performance of BBPSSW for two copies of the input state rho (consisting of two qubits). returns (F_out, p_succ)

EntanglementDist.EPLDParam — Function.

(F, p_succ) = EPLDParam(p, pd)

Determines the performance of EPL-D for rStateCorrPhase. The inputs are the p and pd parameters of rStateCorrPhase. returns (F_out, p_succ)

(F, p_succ) = EPLDParam(rho)

Determines the performance of EPL-D for an arbitrary bipartite state of dimensions nA=nB=4. returns (F_out, p_succ)

EntanglementDist.filtering — Function.

`(rhoQC,P,F,psucc) = filtering(rhoAB,eps)

Implements the filtering protocol on input state rho with filtering parameter eps. This filtering protocol is designed to work well on states of the form p * EPR + (1-p) |10><10|.

Input:

rhoAB 2x2 state to be filtered

eps filtering parameter, smaller eps typically means higher fidelity but lower probability of success

Output:

rhoQC filtered state with flags: Q=hatA, hatB contains the output state, C F_A,F_B contains the flags (1 success, 0 failure)

P projector specifying how Alice and Bob determine success |11><11|, both succeed

F fidelity achieved for the given input state

psucc probability of success

EntanglementDist.filteringMakeChoi — Function.

`(choiA, choiB) = filteringMakeChoi(eps)

Computes the choi states for Alice and Bob in the filtering protocol with filtering parameter eps. This filtering protocol is designed to work well on states of the form p * EPR + (1-p) |10><10|.

Input:

eps filtering parameter, smaller eps typically means higher fidelity but lower probability of success

Output:

choi state for Alice hatA,FA,A

choi state for Bob hatB,FB,B

EntanglementDist.measureScheme — Function.

(rhoQC, F, psucc) = measureScheme(rhoAB, k, MokAlice, MfailAlice, MokBob, MfailBob, fokAlice, ffailAlice, fokBob, ffailBob, P)

Implements a distillation scheme in which Alice and Bob perform the indicated measurement and then output a success or failure flag individually. Global success is determined by the projector P which may compare the to local flag registers.

Input:

rhoAB input state to be distilled

k local dimension of the max entangled state to be produced

MokAlice Kraus operator corresponding to Alice success

MfailAlice Kraus operator corresponding to Alice failure

MokBob Kraus operator corresponding to Bob success

MfailBob Kraus operator corresponding to Bob failure

fokAlice flag Alice outputs for success

ffailAlice flag Alice outputs for failure

fokBob flag Bob outputs for success

ffailBob flag Bob outputs for failure

P projector determining for which joint flags they declare global success

Output:

rhoQC output state where Q=hatA, hatB is the output state and C=F_A,F_B the flag register

F fidelity achieved

psucc success probability achieved

EntanglementDist.measureSchemeMakeChoi — Function.

choiState = measureSchemeMakeChoi(MokAlice, MfailAlice, fokAlice, ffailAlice)

Computes the choi state of a measuring distillation/filtering protocol of one party: Outputs choi_QA where A is a copy of the input and Q is the output of the distillation map A->hatA, fA

Input:

MokAlice Kraus operator corresponding to Alice success

MfailAlice Kraus operator corresponding to Alice failure

fokAlice flag Alice outputs for success

ffailAlice flag Alice outputs for failure