Tools for creating and manipulating quantum states

Tools for creating and manipulating quantum states

Within our package we provide a number of useful functions for dealing with quantum states.

EntanglementDist.isQuantumState — Function.

b = isQuantumState(rho) or b = isQuantumState(rho,prec)

Checks whether the input matrix rho is a valid quantum state, that is, a matrix that is positive semidefinite and has trace 1. prec is the precision that small eigenvalues are dealt with: If the input matrix has a negative eigenvalue that is no smaller than -prec, then this eigenvalue is considered zero.

Returns true/false.

source

EntanglementDist.isHermitian — Function.

b = isHermitian(rho) or b = is Hermitian(rho, prec)

Checks whether the input matrix rho is a Hermitian matrix. Entries are compares up to precision prec on average.

Returns true/false.

source

EntanglementDist.isUnitary — Function.

b = isUnitary(U)

Checks whether the input matrix is unitary.

Returns true/false.

source

EntanglementDist.eVec — Function.

vec = eVec(d, j)

Returns a sparse vector with all 0's except at the position j.

source

EntanglementDist.maxEnt — Function.

vec = maxEnt(d)

Returns a projector (density matrix) onto the maximally entangled state of a given dimension d.

source

EntanglementDist.maxEntVec — Function.

vec = maxEntVec(d)

Returns the maximally entangled state of a given dimension d as a vector.

source

EntanglementDist.copies — Function.

rhoOut = copies(rho,n)

Returns rho^(tensor n), i.e., n copies of the input rho

source

EntanglementDist.sortAB — Function.

rhoSorted = sortAB(rhoBig, d, n)

Given a state of the form rhoBig = rhoAB^(tensor n), rearrange, so all A parts are first, followed by all B parts. d is the dimension of A, assumed to be the same as for B. n is the number of copies.

source

EntanglementDist.entFidelity — Function.

F = entFidelity(rho)

Returns the fidelity with a maximally entangled state of equal dimension.

source

EntanglementDist.eprFidelity — Function.

F = eprFidelity(rho)

For locally two-dimensional states, returns the fidelity with the closest bell state.

source

EntanglementDist.permutesystems — Function.

rhoOut = permutesystems(rho, perm) or rhoOut = permutesystems(rho, perm, dim) or rhoOut = permutesystems(rho,perm, dim, row_only) or rhoOut = permutesystems(rho,perm, dim, row_only, inv_perm)

Permutes the systems in the matrix rho.

Inputs:

rho input matrix

perm a permutation vector on all the systems

dim (optional keyword argument) vector with dimensions of all the systems. By default all systems are of equal dimension.

row_only (optional keyword argument) if set to true, only the rows of rho are permuted.

inv_perm (optional keyword argument) if set to true, the subsystems are permuted according to the inverse of perm rather than perm itself.

Outputs:

rhoOut matrix rho with permuted systems

source

EntanglementDist.partialtrace — Function.

rhoOut = partialtrace(rhoIn, sysNum, dimVec)

Computes the partial trace of rhoIn, tracing out system number sysNum, where the dimensions of the systems are specified in the vector dimVec.

Example: rhoOut = partialtrace(epr, 2, [2 2])

Traces out the second qubit of a 2x2 state, here the EPR pair.

source

EntanglementDist.partialtranspose — Function.

rhoOut = partialtranspose(rho) or rhoOut = partialtranspose(rho, sys) or rhoOut = partialtranspose(rho, sys, dim)

Applies the partial tanspose to the matrix rho.

Inputs:

rho input matrix

systems (optional) single system or a vector of systems to be transposed. By default it is equal to 2. If dim is not provided, the only two possible values of systems are 1 or 2.

dim (optional) vector with dimensions of the systems. By default it assumes two systems of equal dimensions.

Outputs:

rhoOut matrix rho with partial transpose applied

source