Special quantum states

The following functions generate specific sets of quantum states which commonly occur either in physical implemtations or in mathematical studies of entanglement distillation.

EntanglementDist.wernerState — Function.

rho = wernerState(p) or rho = wernerState(p,d)

Returns a werner state, i.e., a mixture of a maximally entangled pair (with probability p) and the maximally mixed state in local dimension d. If no argument d is given the default is d=2, that is, we take the mixture of the EPR pair with the maximally mixed state of local dimension 2.

EntanglementDist.bellDiagState — Function.

rho = bellDiagState(p1,p2,p3)

Returns a bell diagonal state that is a mixture between the 4 Bell states: p1 * phi^+ + p2 * psi^+ + p3 * phi^- + (1 - p1 - p2 - p3) * psi^-.

EntanglementDist.sState — Function.

rho = sState(p)

Returns a state that is an a mixture between the EPR pair (with probability p), and the state |11><11|.

EntanglementDist.rState — Function.

rho = rState(p)

Returns a state that is an a mixture between the EPR pair (with probability p), and the state |01><01|.

EntanglementDist.sStateQutrit — Function.

rho = rStateQutrit(p)

Returns a state that is a mixture between a 3 dimensional maximally entangled state (with probability p) and the state |00><00|

EntanglementDist.rStatePhase — Function.

rho = rStatePhase(p, phi) or rho = rStatePhase(p)

Returns a mixture between a state proportional to |01> + e^(i phi) |10> (with probability p) and the state |11><11|. No value for phi defaults to phi=0.

EntanglementDist.rStateCorrPhase — Function.

rho = rStateCorrPhase(p, pd)

Returns a state of the form integral phi [pd * r(p,phi) + (1-pd) * r(p,phi+pi)] tensor r(p,phi), where r(p,phi) = rStatePhase(p,phi). Default for pd is 1.

EntanglementDist.rStateCorrPhaseCopies — Function.

rho = rStateCorrPhaseCopies(p)

Returns a state of the form integral phi r(p,phi)^(tensor n) , where r(p,phi) = rStatePhase(p,phi) and n is the number of desired copies.